+++ persist/processor/src/test/java/arez/persist/processor/ArezPersistProcessorTest.java
@@ -141,7 +141,7 @@ public final class ArezPersistProcessorTest
   @Test( dataProvider = "compileWithWarnings" )
   public void processCompileWithWarnings( @Nonnull final String classname, @Nonnull final String messageFragment )
   {
-    assertCompilesWithSingleWarning( classname, messageFragment );
+    assertCompilesWithSingleWarningThatCanBeUpgradedToError( classname, messageFragment );
   }
 
   @DataProvider( name = "compileWithoutWarnings" )